+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/item/BWGCreativeTabs.java
@@ -1,6 +1,7 @@
 package net.potionstudios.biomeswevegone.world.item;
 
 import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
@@ -18,12 +19,13 @@ import java.util.function.Supplier;
  * @author Joseph T. McQuigg
  */
 public class BWGCreativeTabs {
-    public static final Supplier<CreativeModeTab> CREATIVE_TAB = createCreativeTab("biomes_weve_gone", () -> BWGItems.BWG_LOGO.get().getDefaultInstance(), BWGBlocks.BLOCK_ITEMS, BWGItems.ITEMS, BWGItems.NO_LANG_ITEMS);
-    public static final Supplier<CreativeModeTab> WOOD_TAB = createCreativeTab("biomes_weve_gone_wood", () -> BWGWood.ASPEN.logstem().asItem().getDefaultInstance(), BWGWood.WOOD_BLOCK_ITEMS);
+    public static final ResourceKey<CreativeModeTab> CREATIVE_TAB = createCreativeTab("biomes_weve_gone", () -> BWGItems.BWG_LOGO.get().getDefaultInstance(), BWGBlocks.BLOCK_ITEMS, BWGItems.ITEMS, BWGItems.NO_LANG_ITEMS);
+    public static final ResourceKey<CreativeModeTab> WOOD_TAB = createCreativeTab("biomes_weve_gone_wood", () -> BWGWood.ASPEN.logstem().asItem().getDefaultInstance(), BWGWood.WOOD_BLOCK_ITEMS);
 
     @SafeVarargs
-    private static Supplier<CreativeModeTab> createCreativeTab(String name, Supplier<ItemStack> icon, ArrayList<Supplier<? extends Item>>... items) {
-        return PlatformHandler.PLATFORM_HANDLER.createCreativeTab(name, icon, items);
+    private static ResourceKey<CreativeModeTab> createCreativeTab(String name, Supplier<ItemStack> icon, ArrayList<Supplier<? extends Item>>... items) {
+        PlatformHandler.PLATFORM_HANDLER.createCreativeTab(name, icon, items);
+        return ResourceKey.create(Registries.CREATIVE_MODE_TAB, BiomesWeveGone.id(name));
     }
 
     public static void tabs() {